+++ src/Forklaringsmodell.Domain/Entities/Vilkar.cs
@@ -11,10 +11,14 @@ public class Vilkar
 {
     public Guid VilkarId { get; set; }
     public string Navn { get; set; } = string.Empty;
+    public string? Kode { get; set; }
+    public string? Kodeverk { get; set; }
     public VirkningType Type { get; set; }
+    public GrunnlagsType Grunnlagstype { get; set; } // rettslig/intern praksis/datakvalitet, se regel 3.15
     public FastsettelsesmateType Fastsettelsesmate { get; set; }
     public string? StandardTekst { get; set; }
     public Guid? RegelId { get; set; }
+    public string? CpsvTjenesteReferanse { get; set; }
 
     public Regel? Regel { get; set; }
     public ICollection<VilkarRettskilde> VilkarRettskilde { get; set; } = new List<VilkarRettskilde>();